Added Client-Side Unity Project
The unity project for the client. There is a bit of server code but it is only for mock data to the functions and scripts that exist. Blender currently is required to view some of the models but it will all be changed to FBX to avoid needing to download Blender on all editing computers.

Unity Version is 2018.1.10f1. Don't update, it might break it, and if someone who isn't on that version edits it, it will keep on switching between versions every commit.