3D Pool Menus Completed
The 3d pool menus have been put in the scene and are fully functional for each resource. I changed some of the base WorldGUI class and other related systems to compensate for some new functionality. Sections for server code exist as prototype functionality but I'll need to make better mock server functions that more accurately depcit what the process will look like. But as far as the client is concerned it won't look much different in terms of functionality.