made Island Management Scene
Updated the Hex Scene so it is now the island management scene. Allows for cycling through islands to view them. I have disabled the orbital cam temporarily because it always needs to have a focus and this system deletes the focus points. I'll add the orbital feature back in later when I add the edit/explore island mode.