Refactoring For Scene Coherence
I've realized that there will need to be some sort of state tracking between scenes. At least to cut down on some of the calls back and forth to the server. Originally I was hoping to just keep it all one scene but more scenes are becoming apparently needed. So I needed to make a class that can track the state between scenes. This required some rewriting and refactoring that was going to probably needed anyways. I tried an overloaded sctruct to avoid the need of a new type of struct but I just caused more problems for myself and messier code so I will fix that next commit. currently still mostly works, but the unit purchasing and resource sends don't.