State Tracking Refactor
Did some state tracking refactoring. I was having some problems trying to keep track of the ste state between the player and the server so I just made a fake server function. It's not actual server request but it does make a nice single point of entry for the code. The rest of the game logic was starting to suffer from complexity issues because server requests were being made through the code and attempted to keep track of the state without calling it directly from the server, which initially I thought was a good idea but turns out not. So basically I'm just going to request the full world,player, and whatever other states, EVERY time a player talks to the server. My assumption is that a call is a call regardless of how big it is within the realms of reasonable JSON file sizes. So in terms of server costs I dont think it adds and costs, but it does reduce the shit out of the complexity and increases the games resillience against crashing and desyncing. All for the same amount of calls I was making in the first place.

So basically I will be converting everything over to that style over the next couple of days. Currently purchasing and pool updates are broken. Don't know why, they should be working, but thats the price of complex tangly code. So I will get that fixed and streamlined.