Fixed State Tracking
State Tracking is fixed as far as I can tell.  A lot of the broken stuff was because I was tring to save bigAmount and amount, bigAmount being ulong for values I knew could only be positive and smaller values that could go negative as amount.  I just removed bigAmount and kept amount as ulong. I figure just do the extra work to hardcode a function meant for specifically subtracting a value rather than a function that accepts a positive or negative value. My thinking is that it should reduce the amount of complexity I'm dealing with.