Created Combat NameSpace
This namespace will hold all of the function structs and classes involved with calculating combat outcomes. This will only be used for the fake server function I have now, but on the server, whether it is c# or javascript, it will do the exact same thing. This will be the actual code needed to calculate real outcomes.