Scene Fully Integrated
Moved all of the Hex stuff successfully over to the IslandMenuScene. Also changed one line under the island menu script for setting the index to zero. Discovered that when overwriting the attackable island you get an out of range index error, so did easy fix and set index to zero. The problem is probably the order of the island count subtraction but I just wanted a working verison  I might redo the island management stuff anyways.