Main Island Preperations
Converting the main island to a functional part right now. I created a squad formation world gui, then I realized I haven't made a world navigator for the new combined scene. Currently the hex island navigator is being used but that does not suit for the new combined scene. So I will have to refactor GameMenu.cs and IslandMenu.cs into one script. I also fixed the UnitManagementUI. One of the problems was coming from the fact the main gameobject was 13.57953 instead of a unit size of 1 so it was screwing with everything. I did not add box colliders to the main island buildings because there were mesh colliders on the child objects. I went to delete them and unity crashed. I'm guessing it has something to do with builderize, so Joshua will have to remove them because I don't know how builderize works and don't want to ruin anything. After all unnecessary meshcolliders are removed from the prefabs, box colliders can be applied to the buildings. Tomorrow a minimum working main island should definetly be done though.

Also, I made a spinning product showcaser for units. I was thinking mayber floating a model above the the appropriate buildings when purchasing units instead of render textures on a gui element.