Click Navigation for MainIsland Implemented.
You can move around the main island with the orbital cam. Still need to fix being able to click other menu revealers on other tiles but it works in basic form right now. Also need to change all of the purchase and pool guis to reflect real names and fix the screen gui to initialize stats on start.