World Navigator Script added.
Made the world navigation script and all of the functions necessary to properly load the interaction scrips with their variables. I decided that the world navigator should just create the scripts then load the values. This will all be done in the loading menu but right now this is done in the scene which shows some lag. It's mostly finished, I just need to fix the moving from island to island, right now it is in explore mode the whole time but I need to turn explore mode on and off so it does the proper motions. I need to clean the tile prefabs up and add the random varation scripts because all of the objects on are making the hex scene very laggy. Testing the discovery scene will be even laggier, so I need to fix that first.