Git pull fix.
I had deleted and repulled to fix some errors on my side, but it broke unity because of packages. It ended up being an easy fix, but just incase we run across it again after deleting and pulling.

In packages/manifest.json, Make sure you change the package manager to 2.0.3. Mine was missing. Changing it fixed it then I updated it. Any others that are broken, change there version to an older one, 1.0.0 being the default lowest, then just manually go into packagemanager under window and manually update them again.

I suspect it is because Library is ignored in the git file, but I have read you should keep it ignored. I don't want to mess with it this far into it, so I will not mess with it. All members should update their packages to the same version before pulling a manifest that says they are updated to a certain version. Because currently it will not update it for you, just the manifest will change.