Collector Purchasing Almost Done
You can't purchase resource collectors yet, but it is almost done. I ran into a slight problem trying to figure out how to get the tile type and the index position in the feature string but that is a small bump, it will be finished next push. But the logic exists for the state master part, but the gui and clicking part doesn't yet.