Added Battle Planners and Some Minor Interaction Tweaks
The classes for doing battle planning logic for defense and attack are done. I just need to work on the interaction script and confirm there are no bugs, then I will work on Xaya code. I might not do too many pushes for a while because I will be doing a lot learning xaya and not too much useful updates for this game until i have a solid grasp. But that will be after my next push.