Made Communication Interface
Made this last night but there was something wrong. I was just connecting in wrong order. You have to Launch GSR before you connect to the XayaDaemon and subscribe to the ZMQ service.

So in order:
 '+' to instantiate the XayaCommander and GameStateRetriever objects.
'Home' to launch the Game State Retriever which initialize the wrapper.
'Insert' to connect the XayaCommander to the XayaDaemon and subscribe the gsr to zmq. This will create a bunch of debug logs via "UpdateBlockProgress"
'Delete' to disconnect.