Created Generic Action Handler
Made some stuff to pull direct block data from the block chain. I would really like a generic handler for the Xaya Communication so I dont have to contantly redo this integration stuff. Might be a bad path, but I'm going to try it anyway. But essentially the the GameStateRetriever hands the json data to a processor that processes it as a generic action first, then we can cycle through the moves data and get it converted to actual usable actions. Which are then processed into valid game states later on.

Right now it's just the Generic Action Parser and the Moves Deserializer. I need to make the method that begins processing it into actual game states.