Done with Self made DB.
So I have decided to stop before I go any further. I feel I have a really good handle on most of the important inner workings libxayagame and their important components. the only thing I am not sure of is how the sqlite databases are being made but I can just run a test on the sqlite storage type and check the format with the sqlite browser website I was using to learn sqlite.

But, I will be stopping the rest of the work on building out the way I was going. It will take the exact same amount of time to build the gsp out either way, but if I continue the way I am going one, I will have to build out a couple of other features and two I will have to redo the gsp in a xaya oriented way anyways. So its best if I just stop now and just work on the gsp like I was originally trying to do but realized I didn't know enough then ended up going this route.

Plus a did a test on the mover memory and the exact same test with my method and mine is way heavier because I had none of it planned out or optimized so I would have to refactor/redo most of it anyways. Which there is no point if I know enough to do it xaya style. So that's what I'm doing.