Lots of Refactoring.
I moved all of the "Needs Converting" into the base Iles of War Folder and added all of the namespaces to the IslesOfWar namespace.

Additionally I refactored pretty much all of them. I refactored based on what the GSP state is going to be. It was mostly structural, but there was some functional refactoring as well. Suprisingly the IslandMenu scene for the most part works. It fails on the start, but if you unpause it works as normal. The reason is I changed some public variables that had been set in the editor so now it is null since I changed it. It's getting late and I don't want to fix those right now. Also I don't need to yet. Going to focus on the GSP functionality for the next week or so. then I can build out and fix the grest of the user interface.

MOST of the functionality exists for the GSP I think. I haven't touched the logic for like 2 or 3 weeks because I have been learning Xaya, so I don't remember everything. But as I build out the GSP it will all start coming back to me. I'll be using the classes I had built for "FakeServer" activity to do the GSP logic. I just had to restructure everything first. So Tomorrow/NextPush I will be building out the first parts of the actual GSP that communicates to the blockchain. :D Super excited to finally get crackin on this part.

After going over all of the stuff today I'm pretty sure it should go pretty fast unless I run into something realizing it is set up kind of messy. But, I can just lift and shift the functionality to whatever other object/method I'm using. So I'm pretty certain this should go pretty fast.