Funtionality for GameStateProcessor Being Built Out
Most of the Game StateProcessor Forward Processing is finished. I just need to make the attacking islands functionality. I know the engagement algorithm is garbage and slow, so I might be tempted to at least half optimize it. I may have no choice but to refactor it anyways considering I have to use random in a certain repeateable way. Which I don't think I was doing in the beginning. Also there are a couple of errors. I left them in there so I remember what needs to be changed in the client side. I also commented some "phase out" lines so I know which structs/classes/and functions will be mostly useless for when im rebuilding the client side visuals and interaction. It will be a lot of "footwork" but not a lot of mental work. so should get done pretty fast. Most of the hard part is over, its just physically getting the ideas down in the right order now and testing for any immediately visible bugs. GETTING CLOSER !!!!!