GSP Forward Processing Finished
Oops, forgot to push lastnight. But the forward processing is finished and I am about to start on the tests to make sure everything will work properly and not break under any condition. This push had some types changed, I'm realizing, probably as I build out these tests, that I should standardize the way I have chosen the types. Its going to lead to some messy complications down the line so I want to head that off before it becomes a bigger problem. I've also added a couple of additional condition checks for the state processor. There are a lot of requirements for each function and each variable in the commands so I have made a couple of functions that handle some of the more complex condition checks.

Everything in the state processor is going to get tests. That logic can not fail. Front end can be buggy, even the xaya commands can submit incorrectly and it will be fine, but the state processor can not be broken.