2 Tests Made and Passing
Made tests for NationUpdate/Add and Search for Islands. I should of got more done today but I was planning a little bit of the type usage I mentioned in last commit and I was highly distracted all day, need to lay off the caffeine I think. But both tests are passing and I beleive are really robust and should catch any possible bugs.

I also ended up adding some functionality to the discover island function because I did not remember that I was only generating them. It was not pulling from any list of already existing ilsands, so I had to figure out how I wanted that to be done.  Right now it just uses the average playerIslandcount to get a undiscovered:owned ratio, but I think I might want to change that up a bit later on. Like get the distance from the mean and modulate an individuals chances or something. I just don't want an explosion of people constantly getting more and more islands either as fodder or resource hording. Might put some other cap on it, but I think island discovery is where to do it. Would save me a shit load of extra condition checking simply by not letting them find so many if they have a lot.