Purchase Unit Testing and Collector Development Testing
I forgot to commit yesterday, so there are two updates in this one. I did the purchasing units tests and the collector development tests. It was going to be develop island in general, but there is ending up being a lot of tests for that function. I had to break it up into collector and defenses. There were some bugs I had to fix in the DevelopIsland function. It was only updating based on the first tile. When it ran across a zero it set the bool to false considering it not developable. So I had to break up defenses and collectors into their own functions. The way I have it now, you could have bad defense parameters but good collector parameters and only the collectors will go through. Which is different from the other simple functions, but I think it is fine. Also, collectors get processed first which means defenses could possibly not have enough resources and not get updated. which is fine, only hackers and buggy clients would be submitting those anyways.

I've also added some extra support functions to clean up the tests a bit. Like the pass fail results add function. That way the string edit logic isn't jumbled in the complex condition checking logic. Just makes it nicer and less code as well.