Expanded and Finished Collector Test
Made some additions to the collector test, there were some cases slipping through that I didn't catch the first time. So fixed that up. Letters work in place of numbers, but I didn't test any of those. They will convert to the same things as the numbers anyways, but just leaving a note here just for future reference. Those ahve not been checked but our client will never send those anyways. In the island.collector part on the client side it gets converted to a number regardless. But the reason that letters work is because it uses the same system as feature creation, so either way A = 0 B = 1 a = 0 b = 1 etc... The the build system converts it to numbers anyways.