Defense Tests Created
Got some defense tests created. Definitely not all, have to test blockers as well. Soemthing is currently wrong with it but I'm too tired to do it right now so I will have to pick it up tomorrow. But the resources in "GetCost" seem not to be generating correctly and they don't seem to be deleting correctly either. Should be an easy fix I fixed most of the defense functions already.