Defense Tests Completed
The blocker and bunker defense tests are now complete. I found a pretty big logical bug, easy fix, but would have missed it if i did not do a test for it. I'm at 81 tests and over 1000 lines now, but with how many tiny easy to miss but devestating bugs I am finding I'll be making sure I'm as rigorous as possible.