Finished Resource Update Tests
All of the variations of Resource Updates is complete. Also changed the style of how resrouces are updated. Instead of how much resources are in the deposit I have changed it to how many extractions are left. This should be much easier as I won't have to worry about over mining or any complicated extraction rates. Just multiply the resource by how many collectors are extracting and decrement 1 for all deposits that have a collector. 121 tests now.