Depleted Resource Submission Tests Complete
The tests for Depleted Resources submission is completed. Had forgotten the function in the StateProcessor so had to add that. Also forgot the resource submission as well. So will be adding that right now and making the tests for it. Should be done tonight. I am hoping to get that and the reward function for both the depleted and the resource pools done tonight as well.