Depleted Reward Function and Testing Done, Resource Reward Function Finished
All of the testing for the Depleted Island Pool is finished. the depleted islands ended up being the warbucks pool instead of chi. Which is fine, solves a couple of problems and increases decentralization. I have also finished the testing for that as well. I have the resource pool reward function finished as well but I have not started testing that one. That should be finished tomorrow. After that I only have the attack cycle to test. And I need to test the state to json and back to state to make sure I didn't fuck up any of the classes. But that will be quick. The attack cycle one will be a bit longer. I have a feeling a couple of days.

It does look like I am going to miss the deadline for 23rd but all of the backend will be done and all I will have to do is modify the front end to handle the new changes. Which should be about a week I think. The idea will be to just read the constantly updated "State" class and don't worry about any other sort of tracking.

Oh and I am missing undo data as well but I will take domob's advice and just use the currentState as the undo data for the private beta. I'll change it mid beta testing just so I can get some experience changing the version in the middle of users using it. Just to see what kind of cluster fucks I will experience. And also it lets me launch the beta sooner.