Attack Testing Completed
All of the main tests are completed. Attack testing turned out to be way simpler than I thought it was going to be. It was the biggest and most complex function but there are only a couple of valid results and possible results are what are tested so naturally it came out to be a similar amount to the other tests. I only have some conformity measure that I need to do. Like checking and making sure all of the variables are proper variables. I initially was trying to be optimal by making longs, ints, uints, floats, etc, but that fucked me as the math started to get more intertwined than I expected so I said fuck it and just started converting everything to doubles. There is also some Random Number generator standardization I need to do and some unit caps for attacks etc. And I haven't done negative value checking yet, but that is trivial, there should be no incoming negative values for any action so I can just check right in the beginning once.