Initial Editting to Connect Frontend To Back
Changed a lot of scripts to begin to connect things to the back end. I'll be doing this through the "ClientInterface" class. Most of the functionality will be lift and shifts from state tracker. Statemaster will be removed eventually and so will state tracker. The client interface needs to keep track of the actual processed gameState and the changes they are making before submitting to the blockchain. These changes are tracked via "queuedActions" and "clientState" on the client interface. Every block update, will recheck to see if the queued actions are valid. If they are not valid anymore, for example an island they had was taken and their defense plan is no longer valid, then you reset the clientState to gamestate and remove the queued data. Or some sort of functionality similar to that. I'm thinking maybe I can just change what is no longer valid instead of ALL of the changes they are doing. But we'll see.

But as of now, the goal is to get all of the old functionality concentrated in client interface and begin making some backend/rpc functionality so we can start actually tracking real game states.