More Progress on Connecting backend to frontend.
Can puchase collectors on islands. The test state is legitimate gameStateProcessor structure so everything will work the same as the blockchain version. I will set the purchase units up next update. Changed some of the tiles as well. There still is one tiny error on the lake islands of a missing hidden object under object revealer script. Forgot to change it but I'll do that next update as well.