Unit Purchases Work
Unit Purchases work on the back to front end. I just realized though, that the IslandMenu might break. I didn't check it but some of the variables won't be set which might mean you will get errors every time you are navigating. I think. Not entirely sure. But just filling the unit purchase related variables in will fix that. So its not a bug just needs to be setup.