Resource Pool GUI connections finished.
The resource pool gui acurately reflects the reflects the percentages and ownership of the pool. I currently do not update the current pool ownership because its extra math that has to be recaulculated all the time and every time a new block comes in those calculations will have to be redone anyways, so I just have the CURRENT percentages that the player owns of the pool and that will be updated AFTER they submit it to the block and it is recalculated with the rest of the updates.