Blocker and Bunker Island Purchasing Done
The blocker and bunker purchase system is done. The blocker purchase prompters icons can be changed up later but right now they are just the same thing as the bunker purchase prompter icons. There are a couple of gui things I want to change later on but they are not important, like when you buy a defense and can't logically build their anymore then it should hide the other prompters immediately instead of having to go to another tile then come back. But just little stuff like that that can be changed.