Attack and Defend Mostly Done
The gui part for Attack and Defending and Island is mostly done now. Was a lot quicker than I thought it was going to be. There was already a battleplan interaction script I had made and forgotten about. I deleted most of it, but it gave me a good idea on how to do it. I did want to get a test squad working moving around on the island but I'm really tired, the computer is lagging really bad again, and it has already crashed unity once on me already so I will just head to bed now. I have a feeling this computer is about to take a shit on me. Anyways, I can probably finish the functionality of the Attack Defend tomorrow and be started or finished on the nation and login stuff as well. GETTING CLOSER BITCHES!!!!