Battle Interaction Complete
All of the main battle interaction is complete. Like everything else, only the commit to queuedAction variable is left. That is only like one line/function for all of them but i wanted to do it at the end. Then after that it is only the submit to block but I need the GSP working for that.