Made Final GUIs
Made the last GUIs for the game as far as I can tell. Might be one that I am not thinking of so not too important if I have forgotten it. But the nation menu and the login menu are done. I also checked out and editted some of the game state and xaya connection stuff. I refamiliarized my self with how it works and how I had it setup. Tomorrow I will get started on being able to connect to the block chain and process blocks. After that I need to start connecting each of the menu parts to actual blockchain functions. I will probably start with finishing the "queuedActions" lines. Then from there is just a matter of converting the queued action to json and updating via XayaCommander script.