Fixed Some StateProcessing stuff.
There were some bugs in the callbacks which was making it a pain in the ass to try and find out why my wrapper was not picking up new changes. So I know now I will by default assume my callbacks have broken it in some manner if the blocks are not updating. I did end up getting it to work, but it crashes at the first encounter of a user. It happens after the first user is added and then on the second block when it starts to Update the Resources it crashes because of a null reference in the player resources list. I will probably make some sort of unit test for a looped version of the callbacks. I have unit tests for all of the individual functions but not for the PlayState() which does the forward processing. So I will make one of those tomorrow as I will likley be editting that section heavily tomorrow to make it work. But the login screen works and processes upto the first encounter of a user move. Which is good.

Sqlite for some reason will not process even the first block, probably just a setting but I dont need to worry about that right now, I'll just get it working in memory first then figure out what is wrong with that. I think I remember having that same problem in the mover demo and fixing it somehow. Might just be a plugin I need to download into the project, I forget. But like I said I will figure that out later.

So fucking pumped this game is nearing its completion.