Backend Fixed
As far as I know the GSP is fixed. I haven't tested all of the commands but it takes too long to handcraft commands and insert them in the testData.json file. But speaking of, the file just consists of an array of string, the string are legit json objects and also legitish block data. I have generated them randomly. Any line that is "x100" or whatever is a string that prompts the system to generated blockdata the amount that comes after the x. The code was useful to check what was going on. When it is running on the chain the errors are obscured because they are on other threads. So it still might come in use down the road as I'm testing everything out.

I'm just going to start connecting everything though. I will make the unit test later on after everything is connected and seemingly playable. It will be a lot easier to create data to test on with an interface that makes legit actions.