Connected Frontend to Backend
The backend communicates successfully with a regtest chain. Everything is coming together boiiiii. Works really well. Wasn't able to get any of the actions via gui implemented, but it does update resources when new blocks are discovered. So it is working. I will definitely have most if not all of the interface working tomorrow. They are just buttons triggering functions that will be about 20 lines or so each. Most of the logic for the commands already exists I just need to make button call functions. There are some other things like recalculating if the user's queued actions are still valid if there is a block update but i'm not too worried about that right now I just need to test all of the different commands. So I will do that tomorrow.