Added Name_Update Functions to GUI
The GUI submits actions on behalf of the players now. And the whole login system works with new players as well. There are only islandmanagement functions like building collectors and bunkers and the battle plans that are left.

Also the submissions submit immediately, that will have to change so players can save money by submitting many actions at once, but I will figure that out in beta. Not necessary right now, the code exists already for making it easy but its not a priority. Just need to make sure all actions are legitimate and processed in the same manner every single time.

Also I added a new random function. Unity's random can't be used outside of the main thread, so had to change all of those. The GSP does use the rng seed now though. There were a lot of little things I did today to many to list, but essentially, just added a bunch of name update functions for the gui. So the user has access to much of the gameplay elements and directly talks to the block chain right now. Only a couple more functions left for it to be fully integrated. That should be done tomorrow I think.