All GUIs Are Connected to their NameUpdates
All of the guis and interactions succesfully complete action creations, with the exception of the battle plans, just need to make a tiny loop to clean any empty arrays out. Forgot to do that, just remembered right now, but the computer is getting laggy and need to restart. Submission will work too I just have to uncomment the SubmitQueuedAction line in either commandInterface or clientInterface I forget which. Pools are not displaying correctly, displaying different values everytime I reopen it, don't know what that is all about, but the unit tests for the resource contribution worked so the backend works. Just need to tidy up the guis, which is what me an joshua will be doing next.