Fixed Some Login Stuff and Added Advanced Options
User can connect with cookies now, modify port numbers, choose wallet, and see messages. There is also a create name window that pops up for wallets that have no names, but it doesn't work because I don't know the XayaService command to RegisterName. It should be RegisterName() according to the documentation but it is not showing up in mine for some reason. The user can also see some log info about connection state now. I will probably delete that later on after full launch, but should help a little in finding problems.

Additionally, the settings have been set to development build with script only building. I don't know if that will effect other people or if it is relative to the computer, but just noting that here.