Secondary Messaging and State size reduction.
I removed variables in the state and playerState that were solely for converting lists to arrays using the .ToArray() method. I hadn't realized it would basically duplicate those variables in the json. I have known about it for a while but thought it was going to be tedious and boring. But turned out to be fairly quick.

I also changed the messaging system to only notify with only one message rather than the entire list. The player actions will be too often and that was annoying seeing it pop up all the time. So only a single message will popup and disappear. But the main message board still exists and the user can open that if they want to review their actions.