Version Checking
This ended up being easier than I thought. I settled with just encapsulating almost the entire ForwardCallback in a simple current version vs client version if statement. On its face it looks like what I want, but I will be rebuilding the unit tests since quite a bit has changed and also I want to test on an actual blockchain rather than just the callbacks. So I will also naturally be testing to see if that works as expected in there, which with how simple it is, I imagine I'm not missing anything. But that's what I am working on next, redoing the unit tests.