Unit Test for the game state processor and AirDrop command.
It's not a unit test like before, where each function is tested. I know that's the point of "unit" testing, but this one is just testing the entire game state processor as a single test and in a blockchain environment, not just from the callbacks. I still need to fix the actual function unit tests, but I wanted to make this one as I haven't had a unit test for the entirety of the process. I have also saved the regtest folder as unitTestRegTest and listed all of the steps I have taken and important checkpoints for the states for debugging purposes if something breaks in a test.

If any of the probabilistic stuff is changed, it will break the test, so make sure you set the probabilities to what they were during when the test was created if you want to test deterministic functionality. Obviously, if any function is overhauled, this test will need to be remade. But I think it is an important test to have considering how important being in sync with other players is.

I have not finished it yet, I need to fix some defend island stuff before I continue with it. For some reason, the defense plans were not submitting and even erased squads but wouldn't let me make a defense plan again. So I am going to fix that.

Also I have added an AirDrop command that takes an array of player names, the amount of resources all of them get in an array of 4 elements and a reason string. The reason string is just so it's a bit more transparent and can add messaging and notifications to the player as well.