Further Development on Unit Test
I've made more progress on the unit test. I've gotten as far as finding an attackable island. Which is why I am committing right now. There will likely be soem problems in attacking that I will need to fix. Also the last problem with the defense was a simple case of >= rather than >. It was allowing me to do it on the GUI but in the actually validation seqeunce, it thought I had too many squads. Probably means I need to make sure they are checking with the same function rather than whatever I did.

I have also changed the searching for islands slightly different. I was thinking about how expensive it could get to continually find attackable islands rather than new ones and that people would probably just make new accounts to get a bunch of islands, which I want to avoid. So instead, I changed the algorithm slightly to only charge you X% of whatever the island search cost is if you discover an attackable island rather than an undiscovered island.  I'm hoping that the smaller cost will be a less of an inconvenience for most players than multiple accounts would be. This variable was also added to the admin commands.

I've deleted the IslandDiscoveryInteraction because that was from way back when and we decided on the current island discovery method a long time ago. Anyways, going to start making some more of the test. Just wanted to push what I have so far before continuing.