Fixing Attack Mode in BattleInteraction
I've been fixing the attack planning in BattleInteraction. It currently needs to load queued attack plans still, that's what I stopped at. But You can load attackable islands and view their defender squads. The entire squads and their counts are loaded as well, so when it comes time to pretty up battle planning, all of the stats are there to use.

I haven't been able to test the following:
Attack Plan Submission to block chain.
Actual Combat loop after submission.
3 Squad limit.

Also, I need to figure out a way to seperate the defender positions from the attacker positions and I need to code "TurnOnDefenses" function for the battle planning. The Turn on defenses should be done tomorrow easily though. The base functions already exist in TileStates, just need to add it to the TurnOnIsland function in the battlehud or whichever script turns on the battle island. Should just be a small function or for loop.