Attacking Works
All of the attack planning, submission and even combat loop work. I don't know if the results of the combat loop are the correct results, but it does process the combat loop and it does remove soliders from both the island and lost units in the attackers, along with any remaining structures on the defeated tile.

I haven't robustly tested any of the combat test loop, I THINK, there might be a problem with the randomness, I thought that I saw three missing machine gunners (as in they died) but now when I reloaded it says that there were 3 riflemen dead. I could be wrong, I was talking to joshua when I ran the loop and wasn't fully paying attention, but I'm pretty sure I saw that the machine gunners lost soldiers not riflemen, so I need to check that and make sure.

So I need to check the combat loop and make sure it is working, and also I need to test and make sure defeating an entire island works as well. Defending the island works and it stays in the hands of the defender if the attacker hasn't killed everyone.