Combat Loop working and Warbucks Port re added
The combat loop seems to be working, As I thought, it wasn't keeping accurate random, but that is to be expected considering I was just using System Random rather than feeding it the seed. So that makes sense. All of the combat stuff uses MudHeroRandom now and references it from the CallBacks stuff where the seed is crated.

I re added the warbucks port as well. The add button is fixed and doesn't give you an error.

I fixed the empty manage island I was getting when I had an attackable island. I had forgotten that originally we were planning making the battle planning in the same island management but, there was too much stuff going on in that menu so I made the battle planning its own island. But I had forgotten to remove the lines of code that were making sure to check for an attackable island in the player object. Good thing it showed up as an empty Island I might not have noticed it otherwise had it shown up as a normal island.