Combat Visualizations
The combat visualization system is almost to its MVP stage. I just need to change the following:

unit sizes.
animations of aircraft and tanks to random idle scripted movement
machine gunner squadron gives a weird index error for some reason with Bass name.
random rotations.

Then it will be useable. There will still be overlapping of squads if I put many squads on the same tile, but its not super important that it is fixed right now. But that can be fixed by tracking squads on tiles in BattlePlanInteraction. On TileStats, where we get spawn position info from, it will be too awkward tracking stuff I think. I'd need a bit of changing in BattlePlanInteraction for that method to work. It's possible, but I'm going to see if there is an easier cleaner method to do before doing that much editting.

As I said though, it is something that is not important right now and I will add that after we have launched open beta.